Moved networking to its own file.
