+++ client/src/main.rs
@@ -1,63 +1,8 @@
-use shared::{ServerRequest, ServerResponse, UserProfile};
-use std::time::Duration;
-use uuid::Uuid;
-use yarws::{Client, Msg};
-
-enum LoginState {
-    LoggedOut,
-    Connected { installation_id: Uuid },
-    Authenticated { profile: UserProfile },
-}
+mod network;
+use network::Network;
 
 fn main() {
     SingleWindowApplication::run(block_on(CosmicCantina::new()));
-}
-
-async fn network_loop() {
-    loop {
-        let mut login_state = LoginState::LoggedOut;
-        let socket = match Client::new("ws://localhost:7878/ws").connect().await {
-            Ok(socket) => socket,
-            Err(err) => {
-                println!("Error connecting to socket. {}", err);
-                tokio::time::delay_for(Duration::from_millis(100)).await;
-                continue;
-            }
-        };
-        let (mut tx, mut rx) = socket.into_channel().await;
-        tx.send(Msg::Binary(
-            bincode::serialize(&ServerRequest::Authenticate {
-                installation_id: None,
-            })
-            .unwrap(),
-        ))
-        .await
-        .unwrap_or_default();
-        while let Some(msg) = rx.recv().await {
-            match msg {
-                Msg::Binary(bytes) => match bincode::deserialize::<ServerResponse>(&bytes) {
-                    Ok(response) => match response {
-                        ServerResponse::Error { message } => {
-                            println!("Authentication error {:?}", message);
-                        }
-                        ServerResponse::AdoptInstallationId { installation_id } => {
-                            println!("Received app token {}", installation_id);
-                            login_state = LoginState::Connected { installation_id };
-                        }
-                        ServerResponse::Authenticated { profile } => {
-                            println!("Authenticated as {}", profile.username);
-                            login_state = LoginState::Authenticated { profile };
-                        }
-                        ServerResponse::AuthenticateAtUrl { url } => {
-                            webbrowser::open(&url).expect("Error launching URL");
-                        }
-                    },
-                    Err(_) => println!("Error deserializing message."),
-                },
-                _ => {}
-            }
-        }
-    }
 }
 
 use futures::executor::block_on;
@@ -146,7 +91,7 @@ impl Window for CosmicCantina {
     }
 
     async fn initialize(&mut self, _scene: &mut Scene) -> KludgineResult<()> {
-        Runtime::spawn(network_loop());
+        Network::spawn().await;
         Ok(())
     }
